pyproject.toml
[tool.poetry]
name = "rcs_pyservice"
version = "0.1.0"
description = ""
authors = ["Mio <liurusi.101@gmail.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.8.0 <4.0"
fastapi = "^0.70.0"
tenacity = "^8.0.1"
aio-pika = "^6.8.0"
uvloop = "^0.16.0"
python-dateutil = "^2.8.2"
pytz = "^2021.3"
tzlocal = "^4.1"
schema = "^0.7.4"
redis = "^4.0.2"
aioredis = "^2.0.0"
loguru = "^0.5.3"
snowland-smx = "^0.3.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
python-multipart = "^0.0.5"
httpx = "^0.21.1"
python-consul2 = "^0.1.5"
dnspython = "^2.1.0"
odmantic = "^0.3.5"
seoservice2020-aiocache = "^1.0.0"

[tool.poetry.dev-dependencies]
#mypy = "^0.770"
#black = "^19.10b0"
#isort = "^4.3.21"
#autoflake = "^1.3.1"
#flake8 = "^3.7.9"
#pytest = "^5.4.1"
#sqlalchemy-stubs = "^0.3"
#pytest-cov = "^2.8.1"

#[tool.isort]
#multi_line_output = 3
#include_trailing_comma = true
#force_grid_wrap = 0
#line_length = 88
#[build-system]
#requires = ["poetry>=0.12"]
#build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
